--- a/packages/grafana-toolkit/src/cli/index.ts
+++ b/packages/grafana-toolkit/src/cli/index.ts
@@ -29,11 +29,13 @@ export const run = (includeInternalScripts = false) => {
       .option('-h, --hot', 'Run front-end with HRM enabled')
       .option('-T, --noTsCheck', 'Run bundler without TS type checking')
       .option('-t, --watchTheme', 'Watch for theme changes and regenerate variables.scss files')
+      .option('--measurePerf', 'Measure webpack performance')
       .description('Starts Grafana front-end in development mode with watch enabled')
       .action(async cmd => {
         await execTask(startTask)({
           watchThemes: cmd.watchTheme,
           noTsCheck: cmd.noTsCheck,
+          measurePerf: !!cmd.measurePerf,
           hot: cmd.hot,
         });
       });