First step to work on webpack performance.

1. Adds https://github.com/stephencookdev/speed-measure-webpack-plugin to allow measurement of webpack plugins and loaders performance
2. Adds `yarn start:perf` script to evaluate `yarn start` performance. This will run the dev build without the `watch` enabled.

---
Baseline:

```
 SMP  ⏱  
General output time took 1 min, 49.38 secs

 SMP  ⏱  Plugins
MonacoEditorWebpackPlugin took 11.49 secs
HotModuleReplacementPlugin took 4.11 secs
HtmlWebpackPlugin took 2.67 secs
ForkTsCheckerWebpackPlugin took 0.403 secs
CleanWebpackPlugin took 0.305 secs
NamedModulesPlugin took 0.044 secs
MiniCssExtractPlugin took 0.012 secs
DefinePlugin took 0 secs

 SMP  ⏱  Loaders
babel-loader took 1 min, 43.88 secs
  module count = 3618
modules with no loaders took 1 min, 38.18 secs
  module count = 2366
css-loader, and 
postcss-loader, and 
sass-loader took 46.88 secs
  module count = 12
expose-loader, and 
expose-loader took 42.44 secs
  module count = 2
mini-css-extract-plugin, and 
css-loader, and 
postcss-loader, and 
sass-loader took 12.52 secs
  module count = 2
monaco-editor-webpack-plugin took 10.69 secs
  module count = 1
css-loader took 9.59 secs
  module count = 39
file-loader took 2.27 secs
  module count = 6
ngtemplate-loader, and 
html-loader took 1.45 secs
  module count = 81
babel-loader, and 
monaco-editor-webpack-plugin, and 
monaco-editor-webpack-plugin took 0.581 secs
  module count = 1
style-loader, and 
css-loader took 0.187 secs
  module count = 39
html-webpack-plugin took 0.034 secs
  module count = 2
```